feat(express): create route file for restoAdmin
